requirements.txt
pandas==2.1.4
pytest==7.4.3
Pillow==10.1.0
pdf2image==1.16.3
img2pdf==0.4.4
PyPDF2==3.0.1
